--- lecture_notes/2019-07-31-kros-ss-2019-note0.md
@@ -1,20 +1,0 @@
-# About This Lecture
-
-이 강의는 로봇학회 여름학교에 참가하신 분들을 대상으로 "Deep Learning based Viusal Odometry and Depth Estimation" (DL-VODE, or simply VODE) 라는 주제로 강의를 합니다. 강의순서는 다음과 같습니다. 
-
-1. **Setup and Train First!**: 뭔지 모르지만 일단 설치하고 학습을 돌려봅니다.
-   1. Setup Colab Instance
-   2. Start Training
-2. **Install Prerequisites in Your Brain**: VODE를 이해하기 위한 사전지식을 공부합니다.
-      1. What is VODE?
-      2. Visual Odometry
-      3. Depth Estimation
-      4. Pose Representation
-3. **Understanding DL-VODE**: VODE의 학습 원리를 이해하고 핵심 과정을 실습합니다.
-   1. DL-VODE structure
-   2. Mathemathical derivation of photometric loss
-   3. Image reconstruction with code
-4. **Paper Review**: SfmLearner 등 최신 VODE 관련 논문을 리뷰합니다.
-5. **Crack the Code**: TF 2.0으로 다시 구현한 SfmLearner 코드를 보며 구현 과정을 이해합니다. 코드를 보기 때문에 강의자료는 따로 없습니다.
-
-
